requirements.txt
Pillow==6.2.1
python-resize-image==1.1.19
